requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

snowflake-snowpark-python
snowflake-snowpark-python[pandas]
snowflake-connector-python
snowflake-sqlalchemy
apache-airflow-providers-snowflake